# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "myArea REST API",
    "description": ""
  },
  "host": "api.pezentics.com",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/api/auth/verify/{token}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}